Add user session key support in CLI frontend
Made several changes, the least of which was swapping out spaces for
tabs. Fixed a bug where `./pb -u username -p' would (incorrectly) keep
going, instead of stopping the program. Added the -k/--key flag, so that
a user can reuse credentials for automation purposes; this led to a
change in the parsing of arguments. When a username is given (with
optional password) without any files to paste, it returns a user key
instead.